+++ backend/filmfinder.py
@@ -1,96 +1,35 @@
 import os
-import requests
 from flask import Flask, request, jsonify, make_response
 from flask_cors import CORS
-from xml.etree import ElementTree
-from models import Theatre, Show
+from finnkinoapi import *
 
 app = Flask(__name__)
 CORS(app)
-
-
-def fetch_location(address):
-    query = {'address': address, 'sensor': False}
-    r = requests.get("https://maps.googleapis.com/maps/api/geocode/json", params=query).json()
-
-    if r['status'] == 'OK':
-        rs = r['results'][0]
-        gm = rs['geometry']
-        location = gm['location']
-    else:
-        location = None
-
-    print(location)
-
-    return location
-
-
-def fetch_cities():
-    """Fetches theatre cities from Finnkino XML API"""
-
-    response = requests.get("http://www.finnkino.fi/xml/TheatreAreas/", timeout=1.000)
-
-    root = ElementTree.fromstring(response.content)
-
-    cities = {}
-
-    for theatre in root.findall('TheatreArea'):
-        area_id = int(theatre.find('ID').text)
-        area_name = theatre.find('Name').text
-        area_name = area_name.split(':')
-
-        # Remove 1029 Valitse teatteri/kaupunki and choose only cities
-        if not area_id == 1029 and len(area_name) == 1:
-            cities[area_id] = area_name[0]
-
-    return cities
-
-
-def fetch_theatres():
-    """Fetches theatres from Finnkino XML API"""
-    response = requests.get("http://www.finnkino.fi/xml/TheatreAreas/", timeout=1.000)
-
-    root = ElementTree.fromstring(response.content)
-
-    theatres = []
-
-    for theatre in root.findall('TheatreArea'):
-        area_id = int(theatre.find('ID').text)
-        area_name = theatre.find('Name').text
-        area_name = area_name.split(':')
-
-        # Remove 1029 Valitse teatteri/kaupunki, remove areas and choose only theatres
-        # TODO: remove hard-coding
-        if area_id == 1029 or area_id == 1014 or area_id == 1012 or area_id == 1002 or area_id == 1021:
-            continue
-
-        if len(area_name) > 1:
-            area_name[1] = area_name[1].lstrip()
-            theatres.append(Theatre(area_id, area_name[1], area_name[0], fetch_location(area_name[0])))
-        else:
-            theatres.append(Theatre(area_id, area_name[0], area_name[0], fetch_location(area_name[0])))
-
-    return theatres
-
-
-def fetch_movie_rating(name):
-    """Fetches movie rating from OMDB API"""
-    query = {'t': name, 'y': 'plot', 'r': 'json'}
-    r = requests.get("http://www.omdbapi.com/", params=query).json()
-
-    if 'imdbRating' in r:
-        rating = r['imdbRating']
-    else:
-        rating = None
-
-    return rating
 
 
 @app.route('/api/cities', methods=['GET'])
 def get_cities():
     cities = fetch_cities()
 
-    return jsonify(cities)
+    return jsonify({'cities': cities})
+
+
+@app.route('/api/cities/<string:city>/theatres', methods=['GET'])
+def get_theatres_in(city):
+    cities = fetch_cities()
+    if city in cities:
+        theatres = fetch_theatres(city)
+        return jsonify({'theatres': [t.serialize() for t in theatres]})
+    return jsonify(None)
+
+
+@app.route('/api/cities/<string:city>/shows', methods=['GET'])
+def get_shows_in(city):
+    cities = fetch_cities()
+    if city in cities:
+        shows = fetch_shows_in(cities[city])
+        return jsonify({'shows': [s.serialize() for s in shows]})
+    return jsonify(None)
 
 
 # returns Theatre -list.
@@ -105,22 +44,7 @@ def get_theatres():
 @app.route('/api/theatres/<int:theatre_id>/shows', methods=['GET'])
 def get_shows(theatre_id):
     # date_arg = request.args['date']
-    query = {'area': theatre_id}  # , 'dt': date_arg}
-    response = requests.get("http://www.finnkino.fi/xml/Schedule/", params=query, timeout=1.000)
-
-    root = ElementTree.fromstring(response.content)
-
-    shows = []
-
-    for show in root.findall('Shows/Show'):
-        show_id = int(show.find('EventID').text)
-        original_title = show.find('OriginalTitle').text
-        title = show.find('Title').text
-        genres = show.find('Genres').text
-        startutc = show.find('dttmShowStartUTC').text
-        length = int(show.find('LengthInMinutes').text)
-        rating = fetch_movie_rating(original_title)
-        shows.append(Show(show_id, title, genres, "Lorem ipsum", length, startutc, rating))
+    shows = fetch_shows_in(theatre_id)
 
     return jsonify({'shows': [s.serialize() for s in shows]})
